Move functional interfaces to com.hazelcast.function
