+++ hazelcast/src/main/java/com/hazelcast/function/BiFunctionEx.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.internal.util.function;
+package com.hazelcast.function;
 
 import com.hazelcast.internal.util.ExceptionUtil;
 
@@ -30,6 +30,8 @@ import static com.hazelcast.internal.util.Preconditions.checkNotNull;
  * @param <T> the type of the first argument to the function
  * @param <U> the type of the second argument to the function
  * @param <R> the type of the result of the function
+ *
+ * @since 4.0
  */
 @FunctionalInterface
 public interface BiFunctionEx<T, U, R> extends BiFunction<T, U, R>, Serializable {
