+++ hazelcast/src/main/java/com/hazelcast/function/ComparatorsEx.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.internal.util.function;
+package com.hazelcast.function;
 
 import com.hazelcast.internal.serialization.BinaryInterface;
