+++ hazelcast/src/main/java/com/hazelcast/function/PredicateEx.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.internal.util.function;
+package com.hazelcast.function;
 
 import com.hazelcast.internal.util.ExceptionUtil;
 
@@ -30,6 +30,8 @@ import static com.hazelcast.internal.util.Preconditions.checkNotNull;
  * java.util.function.Predicate} which declares checked exception.
  *
  * @param <T> the type of the input to the predicate
+ *
+ * @since 4.0
  */
 @FunctionalInterface
 public interface PredicateEx<T> extends Predicate<T>, Serializable {
