+++ hazelcast/src/main/java/com/hazelcast/function/SupplierEx.java
@@ -14,33 +14,36 @@
  * limitations under the License.
  */
 
-package com.hazelcast.internal.util.function;
+package com.hazelcast.function;
 
 import com.hazelcast.internal.util.ExceptionUtil;
 
 import java.io.Serializable;
-import java.util.function.ToIntFunction;
+import java.util.function.Supplier;
 
 /**
- * {@code Serializable} variant of {@link ToIntFunction
- * java.util.function.ToIntFunction} which declares checked exception.
+ * {@code Serializable} variant of {@link Supplier java.util.function.Supplier}
+ * which declares checked exception.
  *
- * @param <T> the type of the input to the function
+ * @param <T> the type of results supplied by this supplier
+ *
+ * @since 4.0
  */
 @FunctionalInterface
-public interface ToIntFunctionEx<T> extends ToIntFunction<T>, Serializable {
+public interface SupplierEx<T> extends Supplier<T>, Serializable {
 
     /**
-     * Exception-declaring version of {@link ToIntFunction#applyAsInt}.
+     * Exception-declaring version of {@link Supplier#get}.
      */
-    int applyAsIntEx(T value) throws Exception;
+    T getEx() throws Exception;
 
     @Override
-    default int applyAsInt(T value) {
+    default T get() {
         try {
-            return applyAsIntEx(value);
+            return getEx();
         } catch (Exception e) {
             throw ExceptionUtil.sneakyThrow(e);
         }
     }
+
 }
